Add new module service-wfs3, move stuff over from service-wfs-client
