+++ service-wfs3/src/main/java/org/oskari/service/wfs3/WFS3Link.java
@@ -1,4 +1,4 @@
-package org.oskari.service.wfs.client;
+package org.oskari.service.wfs3;
 
 import java.util.Objects;
