+++ service-wfs3/src/main/java/org/oskari/service/wfs3/geojson/MapToGeoJSONGeometry.java
@@ -1,4 +1,4 @@
-package org.oskari.service.wfs.client.geojson;
+package org.oskari.service.wfs3.geojson;
 
 import java.util.List;
 import java.util.Map;
