+++ service-wfs3/src/main/java/org/oskari/service/wfs3/geojson/NOPContentHandler.java
@@ -1,4 +1,4 @@
-package org.oskari.service.wfs.client.geojson;
+package org.oskari.service.wfs3.geojson;
 
 import java.io.IOException;
